move akka module to common-lib
