+++ common/src/main/scala/uk/ac/wellcome/finatra/modules/AkkaModule.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.transformer.modules
+package uk.ac.wellcome.finatra.modules
 
 import javax.inject.Singleton
 import uk.ac.wellcome.utils.GuiceAkkaExtension
