+++ ingestor/src/main/scala/uk/ac/wellcome/ingestor/Server.scala
@@ -1,12 +1,12 @@
-package uk.ac.wellcome.platform.transformer
+package uk.ac.wellcome.platform.ingestor
 
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
 import com.twitter.finatra.http.filters.{CommonFilters, LoggingMDCFilter, TraceIdMDCFilter}
 import com.twitter.finatra.http.routing.HttpRouter
 
-import uk.ac.wellcome.platform.transformer.controllers._
-import uk.ac.wellcome.platform.transformer.modules._
+import uk.ac.wellcome.platform.ingestor.controllers._
+import uk.ac.wellcome.platform.ingestor.modules._
 
 object ServerMain extends Server
