+++ transformer/src/main/scala/uk/ac/wellcome/transformer/modules/ActorRegistryModule.scala
@@ -4,6 +4,7 @@ import javax.inject.Singleton
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
 import akka.actor.{ActorRef, ActorSystem, Props}
+import uk.ac.wellcome.finatra.modules.AkkaModule
 import uk.ac.wellcome.platform.transformer.actors._
 
 import uk.ac.wellcome.utils.GuiceAkkaExtension
